// src/types/QBTime.ts
type QBTimeOutput = {
    Response?: string;
    QbId?: string;
    Date?: string;
    Name?: string;
    Hours?: string;
    Minutes?: string;
    StartTime?: string;
    EndTime?: string;
    BreakHours?: string;
    BreakMinutes?: string;
    Description?: string;
    BillableStatus?: string;
    Customer?: string;
    Service?: string;
    CostRate?: string;
    HourlyRate?: string;
    Taxable?: string;
    Class?: string;
    Location?: string;
}

export default QBTimeOutput;